RD progresas
Dirbas su dirbtiniu intelektu ties rašto darbu. Pridėtos mygtukų fukcijos, kuriuos paspaudus galima skaičiuoti lyg skaičiuotuvu. Galima atlikti sudėties, atimties, daugybos, dalybos veiksmus, skaičiuojant skaičius su kableliu ar be kablelio. Programa yra su funkcija, kad skaičiuojant, visi paspausti mygtukai atinkamai būtų atvaizduoti viršuje, kad būtų patogu sekti savo skaičiavimo progresą.
Susidūrus su problemomis ties neteisingomis reiškmėmis ir serverio neišvedimu, pakeičiau int į float x ir y reikšmes, kad visi kintamieji būtų išvedami, nesvarbu ar jie yra su kableliu ar be kablelio.

+++ AV_main.py
@@ -35,8 +35,8 @@ def index():
 @app.route("/skaiciavimas")
 def skaiciuoti():
     operation = request.args.get("zenklai")
-    x = int(request.args.get("x"))
-    y = int(request.args.get("y"))
+    x = float(request.args.get("x"))
+    y = float(request.args.get("y"))
 
     rezult = 0
